media: mc: entity: Extend .has_pad_interdep() operation to expose streams
The media entity .has_pad_interdep() operation returns whether two pads
of an entity are interdependent. The media entity implements the
operation based on knowledge of internal routes and, for entities that
support them (such as V4L2 subdevs), streams. However, streams are not
exposed externally, which results in .has_pad_interdep() returning true
if any stream flows between the two pads. This makes it impossible to
follow a particular stream through an entity.

Fix this by exposing streams to callers. The operation takes an
additional streams parameter that contains the bitmask of streams of
interest on the first pad, and returns the bitmask of interdependent
streams on the second pad.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>