media: xilinx: isp: Drop port count check
The number of ports of the ISP is hardcoded to two in the DT bindings,
and the number of pads to the same value in the driver. The driver still
counts the ports to verify they don't exceed the number of pads. This is
not needed as it has no influence on the driver's operation. Drop the
check.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>